+++ drone_v0/src/library/tranformVar/transformVar.hpp
@@ -3,7 +3,7 @@
 #include <avr/io.h>
 #include <util/delay.h>
 int char_to_int(char *);
-int numeroUnidades(int );
+long numeroUnidades(long );
 void int_to_char(int ,char *,int );
 int int_to_binario(int8_t  numero);
